MAJOR:
Added new scraper for persons ('administration') that had a political mandate but weren't in the Nationalrat or Bundesrat, but rather members of a government/administration (Minister, for instance).
Multiple Additions and clarifications for the other scrapers; also changes the person scraper to now correctly scrape mandates based on the rss-list.
Adapted README as well
Adapted Dashboard to integrate new scraper