Fixes an issue with check_subscriptions because of the missing 'parl_id' field of Debates
Debates didn’t have a parl_id, but we need that as an identifier for
the subscription hashes. This commit adds a property to the Debates
model that generates a Pseudo-ID out of debate type, number and llp (if
applicable), which should be unique.

Furthermore, it adds the parl_id field to the Debates index, to allow
subscription hash generation once more.

This commit needs an index update of the Debates to be run, as well as
a reset of the SubscribedContent hashes. The latter can be achieved
with the following line in a django shell:

```
[content.reset_content_hashes()  for content in
SubscribedContent.objects.all()]
```

Fixes #110