Fixes an issue where latest_mandate wasn't correctly updated for persons
Mandates (and thus party affiliation) for persons weren’t correctly
updated when scraping the detail pages to augment the mandates. This
fixes that issue.

Sadly, we have some rogue mandates, so the fix for this will need a
db-reset (or at least a complete delete of all Mandates and rescanning
of persons, administrations and auditors with -a
ignore_timestamp=True), followed by an index rebuild.

Refers to #71